hands-on-activity/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
plotly
scipy
implicit
umap-learn